module3-nosql-and-document-oriented-databases/module2-sql-for-analysis/requirements.txt
dnspython==2.0.0
numpy==1.19.4
pandas==1.1.4
psycopg2-binary==2.8.6
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
